expose read_amrvac_parfiles at the top level of the frontend
